Add remote carkit to PlayerstateupdateBL.
When voice assistanct has been started we are sending palystaus
changed response as paused, which is a cause of we are not able
hear voice assistant sound over carkit. To avoid this issue add this
remote carkit to PlayerstateupdateBL to not update playstatus as Paused
when Voice assistant has been launched.

CRs-Fixed: 2355834
Change-Id: Id6288ff993ed30fbba3a652f604b14da872e4ed0