Update current playposition correctly for the BLed remote devices as well.
For BLed remote devices we are not updating the current playback state
to the deviceFeatures[deviceIndex].mCurrentPlayState variable when player
paused/stopped. After that when getposition() API has been called as par
of event EVT_PLAY_POS_CHANGED, DUT is not updating the current playpsotion
(i.e.,@pause, instead it is updating the previous playing position). This
fix is to update current playposition correctly for the BLed remote devices
as well by fetching playposition from mCurrentPlayerState which has been
updated with the current state as part of updatePlayerStateAndPosition() API.

CRs-Fixed: 2408125
Change-Id: I3b699f9ff9f2cba574ca3194572030dc8e8ce757